+++ app/notifications/notification.py
@@ -11,15 +11,15 @@ import pytz
 from apprise import Apprise, AppriseAttachment, NotifyFormat
 from geopy import distance, point
 
-from . import geocoding
+from ..geocoding import geocoding
 from .config import (
     AlertConfig,
     NotificationConfig,
     get_notifications_config,
-    get_ttl_hash,
 )
-from .metadata import Metadata
-from .transcript import Transcript
+from ..models.metadata import Metadata
+from ..models.transcript import Transcript
+from ..utils.cache import get_ttl_hash
 
 
 # TODO: write tests
