+++ app/radio/digital.py
@@ -1,10 +1,10 @@
 import os
 from threading import Lock
 
-from .metadata import Metadata, SrcListItem
-from .transcript import Transcript
-from .whisper.base import WhisperSegment, WhisperResult
-from .whisper.transcribe import transcribe
+from ..models.metadata import Metadata, SrcListItem
+from ..models.transcript import Transcript
+from ..whisper.base import WhisperSegment, WhisperResult
+from ..whisper.transcribe import transcribe
 
 
 def get_closest_src(srcList: list[SrcListItem], segment: WhisperSegment):
