+++ app/whisper/exceptions.py
@@ -1,0 +1,5 @@
+from ..utils.exceptions import BaseException
+
+
+class WhisperException(BaseException):
+    ...
